Fix incorrect reference to curve order
The length of `d` is defined by the order of the largest prime order subgroup,
not necessarily the order of the curve (these coincide when the curve is of
prime order, but these BLS curves are not of prime order). Or, equivalently, the
order of the generator point chosen for that subgroup.

Also change `n` to `r` to align with the notation in pairing-friendly-curves:
https://datatracker.ietf.org/doc/html/draft-irtf-cfrg-pairing-friendly-curves-12#name-selection-of-pairing-friend